Move type-resolution family out-of-line into MLIRGenTypes.cpp (review doc 1)
Second slice of the TU split: the contiguous type-resolution family
(getType ... getNullType, 134 methods, ~3,400 lines) moves from inline
bodies in MLIRGenImpl.h to out-of-line definitions in MLIRGenTypes.cpp.
The lone template member (getIndexedAccessTypeForArray<T>) stays inline
in the header. Default arguments remain on declarations only.

Verified: MLIRGenTests 12/12, full ctest 681/681 (JIT + AOT).
